selinux.linux-pam: admit 1.7.1
Since the introduction of linux-pam 1.7.1 in nixpkgs 25.11 in 26d61f,
the previously automatic SELinux support was explicitly disabled.
This change makes it automatic again, building SELinux support depending
on the presence of libselinux during the build

Additionally, localstatedir now has to be writable, and thus had to
change from the default /var to /tmp, allowing the build to create the
empty directory ${localstatedir}/run/sepermit.